+++ graphify/extract/_dispatch.py
@@ -1,0 +1,218 @@
+"""Main extract() and collect_files() entry points."""
+from __future__ import annotations
+
+import os
+from pathlib import Path
+from typing import Any
+
+from ..cache import load_cached, save_cached
+from ._base import _make_id
+from ._configs import (
+    PYTHON_CONFIG, JS_CONFIG, TS_CONFIG, JAVA_CONFIG,
+    C_CONFIG, CPP_CONFIG, RUBY_CONFIG, CSHARP_CONFIG,
+    KOTLIN_CONFIG, SCALA_CONFIG, PHP_CONFIG, LUA_CONFIG, SWIFT_CONFIG,
+)
+from ._generic import _extract_generic
+from ._elixir import extract_elixir
+from ._go import extract_go
+from ._objc import extract_objc
+from ._rust import extract_rust
+from ._zig import extract_zig
+from ._powershell import extract_powershell
+from ._python_extra import _extract_python_rationale, _resolve_cross_file_imports
+
+
+# ── Public extractor functions ───────────────────────────────────────────────
+
+def extract_python(path: Path) -> dict:
+    """Extract classes, functions, and imports from a .py file via tree-sitter AST."""
+    result = _extract_generic(path, PYTHON_CONFIG)
+    if "error" not in result:
+        _extract_python_rationale(path, result)
+    return result
+
+
+def extract_js(path: Path) -> dict:
+    """Extract classes, functions, arrow functions, and imports from a .js/.ts/.tsx file."""
+    config = TS_CONFIG if path.suffix in (".ts", ".tsx") else JS_CONFIG
+    return _extract_generic(path, config)
+
+
+def extract_java(path: Path) -> dict:
+    """Extract classes, interfaces, methods, constructors, and imports from a .java file."""
+    return _extract_generic(path, JAVA_CONFIG)
+
+
+def extract_c(path: Path) -> dict:
+    """Extract functions and includes from a .c/.h file."""
+    return _extract_generic(path, C_CONFIG)
+
+
+def extract_cpp(path: Path) -> dict:
+    """Extract functions, classes, and includes from a .cpp/.cc/.cxx/.hpp file."""
+    return _extract_generic(path, CPP_CONFIG)
+
+
+def extract_ruby(path: Path) -> dict:
+    """Extract classes, methods, singleton methods, and calls from a .rb file."""
+    return _extract_generic(path, RUBY_CONFIG)
+
+
+def extract_csharp(path: Path) -> dict:
+    """Extract classes, interfaces, methods, namespaces, and usings from a .cs file."""
+    return _extract_generic(path, CSHARP_CONFIG)
+
+
+def extract_kotlin(path: Path) -> dict:
+    """Extract classes, objects, functions, and imports from a .kt/.kts file."""
+    return _extract_generic(path, KOTLIN_CONFIG)
+
+
+def extract_scala(path: Path) -> dict:
+    """Extract classes, objects, functions, and imports from a .scala file."""
+    return _extract_generic(path, SCALA_CONFIG)
+
+
+def extract_php(path: Path) -> dict:
+    """Extract classes, functions, methods, namespace uses, and calls from a .php file."""
+    return _extract_generic(path, PHP_CONFIG)
+
+
+def extract_lua(path: Path) -> dict:
+    """Extract functions, methods, require() imports, and calls from a .lua file."""
+    return _extract_generic(path, LUA_CONFIG)
+
+
+def extract_swift(path: Path) -> dict:
+    """Extract classes, structs, protocols, functions, imports, and calls from a .swift file."""
+    return _extract_generic(path, SWIFT_CONFIG)
+
+
+# ── Dispatch table ───────────────────────────────────────────────────────────
+
+_DISPATCH: dict[str, Any] = {
+    ".py": extract_python,
+    ".js": extract_js,
+    ".ts": extract_js,
+    ".tsx": extract_js,
+    ".go": extract_go,
+    ".rs": extract_rust,
+    ".java": extract_java,
+    ".c": extract_c,
+    ".h": extract_c,
+    ".cpp": extract_cpp,
+    ".cc": extract_cpp,
+    ".cxx": extract_cpp,
+    ".hpp": extract_cpp,
+    ".rb": extract_ruby,
+    ".cs": extract_csharp,
+    ".kt": extract_kotlin,
+    ".kts": extract_kotlin,
+    ".scala": extract_scala,
+    ".php": extract_php,
+    ".swift": extract_swift,
+    ".lua": extract_lua,
+    ".toc": extract_lua,
+    ".zig": extract_zig,
+    ".ps1": extract_powershell,
+    ".ex": extract_elixir,
+    ".exs": extract_elixir,
+    ".m": extract_objc,
+    ".mm": extract_objc,
+}
+
+
+def extract(paths: list[Path]) -> dict:
+    """Extract AST nodes and edges from a list of code files.
+
+    Two-pass process:
+    1. Per-file structural extraction (classes, functions, imports)
+    2. Cross-file import resolution: turns file-level imports into
+       class-level INFERRED edges (DigestAuth --uses--> Response)
+    """
+    per_file: list[dict] = []
+
+    # Infer a common root for cache keys
+    try:
+        if not paths:
+            root = Path(".")
+        elif len(paths) == 1:
+            root = paths[0].parent
+        else:
+            common_len = sum(
+                1 for i in range(min(len(p.parts) for p in paths))
+                if len({p.parts[i] for p in paths}) == 1
+            )
+            root = Path(*paths[0].parts[:common_len]) if common_len else Path(".")
+    except Exception:
+        root = Path(".")
+
+    for path in paths:
+        extractor = _DISPATCH.get(path.suffix)
+        if extractor is None:
+            continue
+        cached = load_cached(path, root)
+        if cached is not None:
+            per_file.append(cached)
+            continue
+        result = extractor(path)
+        if "error" not in result:
+            save_cached(path, result, root)
+        per_file.append(result)
+
+    all_nodes: list[dict] = []
+    all_edges: list[dict] = []
+    for result in per_file:
+        all_nodes.extend(result.get("nodes", []))
+        all_edges.extend(result.get("edges", []))
+
+    # Add cross-file class-level edges (Python only)
+    py_paths = [p for p in paths if p.suffix == ".py"]
+    py_results = [r for r, p in zip(per_file, paths) if p.suffix == ".py"]
+    cross_file_edges = _resolve_cross_file_imports(py_results, py_paths)
+    all_edges.extend(cross_file_edges)
+
+    return {
+        "nodes": all_nodes,
+        "edges": all_edges,
+        "input_tokens": 0,
+        "output_tokens": 0,
+    }
+
+
+def collect_files(target: Path, *, follow_symlinks: bool = False) -> list[Path]:
+    if target.is_file():
+        return [target]
+    _EXTENSIONS = {
+        ".py", ".js", ".ts", ".tsx", ".go", ".rs",
+        ".java", ".c", ".h", ".cpp", ".cc", ".cxx", ".hpp",
+        ".rb", ".cs", ".kt", ".kts", ".scala", ".php", ".swift",
+        ".lua", ".toc", ".zig", ".ps1",
+        ".m", ".mm",
+    }
+    if not follow_symlinks:
+        results: list[Path] = []
+        for ext in sorted(_EXTENSIONS):
+            results.extend(
+                p for p in target.rglob(f"*{ext}")
+                if not any(part.startswith(".") for part in p.parts)
+            )
+        return sorted(results)
+    # Walk with symlink following + cycle detection
+    results = []
+    for dirpath, dirnames, filenames in os.walk(target, followlinks=True):
+        if os.path.islink(dirpath):
+            real = os.path.realpath(dirpath)
+            parent_real = os.path.realpath(os.path.dirname(dirpath))
+            if parent_real == real or parent_real.startswith(real + os.sep):
+                dirnames.clear()
+                continue
+        dp = Path(dirpath)
+        if any(part.startswith(".") for part in dp.parts):
+            dirnames.clear()
+            continue
+        for fname in filenames:
+            p = dp / fname
+            if p.suffix in _EXTENSIONS and not fname.startswith("."):
+                results.append(p)
+    return sorted(results)
